backend/test_gemini_models.py
import os
from google import genai
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("GEMINI_API_KEY")
client = genai.Client(api_key=api_key)

print(f"Testing API Key: {api_key[:5]}...{api_key[-5:]}")

try:
    print("\nAvailable models:")
    for m in client.models.list():
        print(f"- {m.name}")
except Exception as e:
    print(f"\nError listing models: {e}")
