backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
atlassian-python-api==4.0.7
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
Deprecated==1.3.1
distro==1.9.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-core==1.3.3
langchain-protocol==0.0.15
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.2
numpy==2.4.4
oauthlib==3.3.1
openai==2.35.0
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pgvector==0.4.2
psycopg2-binary==2.9.12
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
wrapt==2.1.2
xxhash==3.7.0
zstandard==0.25.0
defusedxml==0.7.1
jira==3.10.5

aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
dataclasses-json==0.6.7
frozenlist==1.8.0
httpx-sse==0.4.3
langchain-classic==1.0.6
langchain-community==0.4.1
langchain-text-splitters==1.1.2
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
ollama==0.6.2
propcache==0.4.1
pydantic-settings==2.14.0
typing-inspect==0.9.0
yarl==1.23.0

langchain-ollama==1.1.0

cuda-bindings==13.2.0
cuda-pathfinder==1.5.4
cuda-toolkit==13.0.2
filelock==3.29.0
fsspec==2026.4.0
hf-xet==1.5.0
huggingface_hub==1.14.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
Pygments==2.20.0
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
transformers==5.8.0
triton==3.6.0
typer==0.25.1
anthropic==0.101.0
cffi==2.0.0
cryptography==48.0.0
docstring_parser==0.18.0
filetype==1.2.0
google-auth==2.52.0
google-genai==1.75.0
GPUtil==1.4.0
langchain-anthropic==1.4.3
langchain-google-genai==4.2.2
langchain-openai==1.2.1
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
tiktoken==0.12.0
websockets==16.0